Add QuicPacketCreator and QuicFecGroup.
Review URL: https://chromiumcodereview.appspot.com/11238013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@163724 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/quic/quic_utils.cc
@@ -10,7 +10,7 @@
 namespace net {
 
 // static
-int QuicUtils::StreamFramePacketOverhead(int num_frames) {
+size_t QuicUtils::StreamFramePacketOverhead(int num_frames) {
   // TODO(jar): Use sizeof(some name).
   return kPacketHeaderSize +
          1 +   // frame count
